Dedublicated code in tests, reorganized tests into folder
