+++ tests/vector tests/test_vector_exceptions.py
