+++ tests/vector tests/test_vector_scalar_product.py
@@ -23,3 +23,20 @@ def test_vector_scalar_product_opposite_directions():
     assert (
         Vector.scalar_product(vector5, vector6) == 0
     ), "Скалярное произведение [-1, 1] и [1, 1] должно быть 0"
+
+def test_scalar_product_of_zero_vectors():
+    vector1 = Vector([0, 0])
+    vector2 = Vector([0, 0])
+
+    assert (
+        Vector.scalar_product(vector1, vector2) == 0
+    ), "Скалярное произведение нулевых векторов должно быть 0"
+
+def test_scalar_product_with_zero_vector():
+    vector3 = Vector([1, 1])
+    vector4 = Vector([0, 0])
+
+    assert (
+        Vector.scalar_product(vector3, vector4) == 0
+    ), "Скалярное произведение любого вектора с нулевым должно быть 0"
+
